update: put ws commands on seperate file also increase agent timeout
